Fix instabillity in L-SVRG
The stage-based approach used seem to have some stability issues. It's
probably due to the sampling from the geometric distribution. Revert
back to the weighted coin-toss each iterations. Wanted to avoid this to
begin with mainly to avoid another branch-there just because. However,
does not affect much, it is already a branch there.